Reorganization of Shiro related files
