+++ src/main/java/org/ohdsi/webapi/shiro/filters/InvalidateAccessTokenFilter.java
@@ -1,10 +1,11 @@
-package org.ohdsi.webapi.shiro;
+package org.ohdsi.webapi.shiro.filters;
 
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletResponse;
 import org.apache.shiro.web.servlet.AdviceFilter;
 import org.apache.shiro.web.util.WebUtils;
+import org.ohdsi.webapi.shiro.TokenManager;
 
 /**
  *
