+++ src/main/java/org/ohdsi/webapi/shiro/filters/SendTokenInRedirectFilter.java
@@ -16,7 +16,7 @@
  * Authors: Mikhail Mironov
  *
  */
-package org.ohdsi.webapi.shiro;
+package org.ohdsi.webapi.shiro.filters;
 
 import org.apache.shiro.web.servlet.AdviceFilter;
 import org.springframework.beans.factory.annotation.Value;
